fix couple issues found when compiling on 32-bit OS

+++ utils/diodls.c
@@ -203,12 +203,12 @@ lsfile_l (Npcfid *dir, char *name)
     if (!(gr = getgrgid (sb.st_gid)))
         snprintf (gid, sizeof (gid), "%d", sb.st_gid);
     mtime = ctime( &sb.st_mtime);
-    printf ("%10s %4lu %s %s %12lu %.*s %s\n",
+    printf ("%10s %4lu %s %s %12ju %.*s %s\n",
             mode2str (sb.st_mode),
-            sb.st_nlink,
+            (unsigned long)sb.st_nlink,
             pw ? pw->pw_name : uid,
             gr ? gr->gr_name : gid,
-            sb.st_size,
+            (uintmax_t)sb.st_size,
             (int)strlen (mtime) - 13, mtime + 4,
             name);
     return;
